# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gestion Locative API
  description: |
    API REST pour l'application SaaS de gestion locative.
    
    ## Authentication
    L'API utilise JWT Bearer tokens. Incluez le header `Authorization: Bearer <token>` dans vos requêtes.
    
    ## Idempotency
    Pour les opérations POST qui modifient l'état (paiements, signatures), incluez le header `Idempotency-Key: <uuid>`.
    
    ## Rôles
    - **admin**: Accès complet
    - **owner**: Gestion de ses propriétés
    - **tenant**: Gestion de son bail et paiements
    - **provider**: Gestion de ses interventions
  version: 1.0.0
  contact:
    name: Support
    email: support@gestionlocative.fr
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /leases/{lid}/rent-invoices:
    post:
      tags: [Invoices]
      summary: Émettre une facture de loyer
      security:
        - bearerAuth: []
      parameters:
        - name: lid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Facture créée
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...